+++ pyogrio/raw.py
@@ -369,9 +369,27 @@ def open_arrow(
 
     See docstring of `read` for parameters.
 
-    The RecordBatchStreamReader is reading from a stream provided by OGR and must not be
+    The RecordBatchReader is reading from a stream provided by OGR and must not be
     accessed after the OGR dataset has been closed, i.e. after the context manager has
     been closed.
+
+    By default this function returns a `pyarrow.RecordBatchReader`. Optionally,
+    you can use this function without a `pyarrow` dependency by specifying
+    ``return_pyarrow=False``. In that case, the returned reader will be a
+    generic object implementing the `Arrow PyCapsule Protocol`_ (i.e. having
+    an `__arrow_c_stream__` method). This object can then be consumed by
+    your Arrow implementation of choice that supports this protocol.
+
+    .. _Arrow PyCapsule Protocol: https://arrow.apache.org/docs/format/CDataInterface/PyCapsuleInterface.html
+
+    Other Parameters
+    ----------------
+    batch_size : int (default: 65_536)
+        Maximum number of features to retrieve in a batch.
+    return_pyarrow : bool (default: True)
+        If False, return a generic ArrowStream object instead of a pyarrow
+        RecordBatchReader. This object needs to be passed to another library
+        supporting the Arrow PyCapsule Protocol to consume the stream of data.
 
     Examples
     --------
@@ -385,12 +403,22 @@ def open_arrow(
     >>>     for table in reader:
     >>>         geometries = shapely.from_wkb(table[meta["geometry_name"]])
 
+    Or without directly returning a pyarrow object:
+
+    >>> with open_arrow(path) as source:
+    >>>     meta, stream = source
+    >>>     reader = pa.RecordBatchReader.from_stream(stream)
+    >>>     for table in reader:
+    >>>         geometries = shapely.from_wkb(table[meta["geometry_name"]])
+
     Returns
     -------
-    (dict, pyarrow.RecordBatchStreamReader)
+    (dict, pyarrow.RecordBatchReader or ArrowStream)
 
         Returns a tuple of meta information about the data source in a dict,
-        and a pyarrow RecordBatchStreamReader with data.
+        and a data stream object (a pyarrow RecordBatchReader if
+        `return_pyarrow` is set to True, otherwise a generic ArrowStrem
+        object).
 
         Meta is: {
             "crs": "<crs>",