+++ CHANGES.md
@@ -7,6 +7,11 @@
 -   `read_arrow` and `open_arrow` now provide
     [GeoArrow-compliant extension metadata](https://geoarrow.org/extension-types.html),
     including the CRS, when using GDAL 3.8 or higher (#366).
+-   The `open_arrow` function can now be used without a `pyarrow` dependency. In
+    that case, specify `use_pyarrow=False` and the returned reader will be a
+    generic object implementing the [Arrow PyCapsule Protocol](https://arrow.apache.org/docs/format/CDataInterface/PyCapsuleInterface.html) (i.e. having an `__arrow_c_stream__`
+    method). This object can then be consumed by your Arrow implementation of choice
+    that supports this protocol (#349).
 -   Warn when reading from a multilayer file without specifying a layer (#362).